Fixed picking. Added distance fog (mainly to prove to myself I can, will probably remove later) Turned off culling based on distance from camera. Will have to implement level of detail / quad tree to speed it up.

+++ ControlManager.cpp
@@ -60,7 +60,7 @@ namespace game_utils
 			GLfloat delta = CV_GAME_MANAGER->getDeltaTime();
 			
 			//zoom with scroll wheel
-			short scroll = -input.checkScroll()/120;
+			float scroll = -(float)input.checkScroll()/120.0f;
 			camera.move(vector3f(0.0f,camZoomSpeed*delta*scroll,0.0f));					
 
 			//clip camera pos
